Update: Decoupling the worker module
Extending the process and worker modules
Feat: Coroutine added context extend
Update: stream::waitForxxable timeout supports float
Project: Update logo and .tarignore
Update: stream::waitForxxable timeout supports float
Update: Refactor parallel && for Windows compatibility
Feat: Worker support metadata

+++ src/Proc/Session.php
@@ -13,6 +13,7 @@
 namespace Ripple\Proc;
 
 use Closure;
+use Ripple\Kernel;
 use Ripple\Stream\Exception\ConnectionException;
 use Ripple\Utils\Output;
 use Throwable;
@@ -198,6 +199,11 @@ class Session
      */
     public function inputSignal(int $signalCode): bool
     {
+        /*** @compatible:Windows */
+        if (!Kernel::getInstance()->supportProcessControl()) {
+            return false;
+        }
+
         return posix_kill($this->getStatus('pid'), $signalCode);
     }
